+++ src/components/server/ServerOverview.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="server-container">
-    <channel-list :channels="this.server.channels" :user="this.user" @change-channel="changeChannel" class="channel-list" />
+    <channel-list :channels="this.server.channels" @change-channel="changeChannel" class="channel-list" />
     <channel-chat :history="this.channel.history" :members="this.members" class="channel-chat" />
     <channel-members :members="this.members" class="channel-members" />
   </div>
@@ -10,20 +10,16 @@
 import ChannelList from './channel/ChannelList.vue';
 import ChannelChat from './channel/ChannelChat.vue';
 import ChannelMembers from './channel/ChannelMembers.vue';
-import {getMembers} from '../../helpers/helpers.js';
+import axios from 'axios';
 
 export default {
   props: {
-    server: {required: true, type: Object},
-    user: {required: true, type: Object}
+    server: {required: true, type: Object}
   },
   watch: {
-    server: function(newServer) {
-      if (newServer) {
-        this.members = getMembers(newServer.members);
-        this.channel = newServer.channels[0];
-      }
-      this.server = newServer;
+    server: {
+      immediate: true,
+      handler: "serverChanged"
     }
   },
   components: {
@@ -40,6 +36,19 @@ export default {
   methods: {
     changeChannel(channelId) {
       this.channel = this.server.channels.find(c => c.cid === channelId);
+    },
+    serverChanged(newServer) {
+      axios
+        .post("/user/members/", {members: newServer.members, server: newServer.name})
+        .then(res => {
+          if (res.status === 200) {
+            this.members = res.data;
+            this.channel = newServer.channels[0];
+          }
+        })
+        .catch(err => console.error(err));
+
+      this.server = newServer;
     }
   }